move subscriber into submodule
