+++ docs/reference/channels/index.rst
@@ -4,4 +4,5 @@ channels
 .. toctree::
 
     plugin
+    subscriber
     backends/index
