+++ docs/reference/channels/plugin.rst
@@ -3,6 +3,4 @@ plugin
 
 .. autoclass:: litestar.channels.plugin.ChannelsPlugin
 
-.. autoclass:: litestar.channels.plugin.Subscriber
-
 .. autoclass:: litestar.channels.plugin.ChannelsException
