+++ docs/reference/channels/subscriber.rst
@@ -1,0 +1,4 @@
+subscriber
+==========
+
+.. autoclass:: litestar.channels.subscriber.Subscriber
